requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
pygments==2.16.1
pydantic==2.4.2
python-jose==3.3.0
bcrypt==4.0.1
sqlalchemy==2.0.23
python-dotenv==1.0.0